tests/wpt/tests/intersection-observer/v2/3d-transform-z-order-occlusion.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #container {
    position: relative;
    transform: translateZ(-10px);
    transform-style: preserve-3d;
    width: 200px;
    height: 200px;
    background: blue;
  }
  #target {
    position: absolute;
    left: 0;
    top: 0;
    width: 100px;
    height: 100px;
    background: red;
  }
  #occluder {
    position: absolute;
    left: 0;
    top: 0;
    width: 100px;
    height: 100px;
    transform: translateZ(5px);
    background: green;
    pointer-events: none;
  }
</style>
<div id="container">
  <div id="target"></div>
  <div id="occluder"></div>
</div>

<script>
  setup({ single_test: true });
  const target = document.getElementById("target");
  new IntersectionObserver(entries => {
    assert_false(entries[0].isVisible);
    done();
  }, {trackVisibility: true, delay: 100}).observe(target);
</script>
